Add query parameters to get measurement by ID
Adds support for include the query parameters exclude_coordinates and contentType to the
function to fetch measurement by ID.

If the content type of the measurement to fetch is known, providing the
content type will speed up the request. If the empty string is passed
the query parameter is excluded.

Setting excludeCoordinates to true will include the corresponding query
parameter, which makes the Measurements service exclude coordinates from
the returned measurement. If the coordinates are not needed this can
significantly reduce the size of a returned measurement.