Add function for verbose post node data (#20)
The verbose query parameter is optional when posting node data. Setting it to true will result in 202 Accepted being returned with a response body containing the measurement IDs of the ingested measurements.

This is mostly used in testing scenarios when the measurement ID is needed for further test verification.

Due to how data ingestion works in the API the returned measurement ID is not guaranteed to exists. Clients should not rely on this in any production use cases.